fix(dashboards): address 61746 review on widgetFilters scope and MCP copy
