+++ products/dashboards/backend/widgets/widget_filters.py
@@ -98,79 +98,6 @@ def validate_widget_filters(config: dict[str, Any]) -> dict[str, dict[str, Any]]
     return validated or None
 
 
-# Per-widget-type event property allowlists for config.widgetFilters (not dashboard quick_filter_ids).
-# Keep in sync with ERROR_TRACKING_WIDGET_FILTER_PROPERTY_NAMES in
-# products/dashboards/frontend/widgets/error_tracking/constants.ts when that file lands.
-ERROR_TRACKING_WIDGET_FILTER_PROPERTY_NAMES = frozenset(
-    {
-        "$environment",
-        "$current_url",
-        "$pathname",
-        "$team",
-        "$posthog_team",
-        "$temporal_worker",
-        "$temporal_worker_name",
-    }
-)
-
-# Keep in sync with SESSION_REPLAY_WIDGET_FILTER_PROPERTY_NAMES in
-# products/dashboards/frontend/widgets/session_replay/constants.ts
-SESSION_REPLAY_WIDGET_FILTER_PROPERTY_NAMES = frozenset(
-    {
-        "$browser",
-        "$os",
-        "$device_type",
-        "$geoip_country_code",
-        "$geoip_city_name",
-        "$current_url",
-        "$pathname",
-        "$host",
-        "$referring_domain",
-        "$lib",
-        "$environment",
-    }
-)
-
-
-def _validate_widget_filters_allowlist(
-    config: dict[str, Any],
-    allowed_property_names: frozenset[str],
-    widget_type_label: str,
-) -> dict[str, dict[str, Any]] | None:
-    validated = validate_widget_filters(config)
-    if validated is None:
-        return None
-
-    for filter_id, entry in validated.items():
-        property_name = entry["propertyName"].strip().lower()
-        if property_name not in allowed_property_names:
-            raise DRFValidationError(
-                {
-                    "config": (
-                        f"widgetFilters.{filter_id} uses unsupported property "
-                        f"{entry['propertyName']!r} for {widget_type_label}."
-                    )
-                }
-            )
-    return validated
-
-
-def validate_session_replay_widget_filters(config: dict[str, Any]) -> dict[str, dict[str, Any]] | None:
-    return _validate_widget_filters_allowlist(
-        config,
-        SESSION_REPLAY_WIDGET_FILTER_PROPERTY_NAMES,
-        "session replay list widgets",
-    )
-
-
-def validate_error_tracking_widget_filters(config: dict[str, Any]) -> dict[str, dict[str, Any]] | None:
-    return _validate_widget_filters_allowlist(
-        config,
-        ERROR_TRACKING_WIDGET_FILTER_PROPERTY_NAMES,
-        "error tracking list widgets",
-    )
-
-
 def _event_property_filters_from_widget_filters(
     widget_filters: dict[str, dict[str, Any]],
 ) -> list[dict[str, Any]]:
