+++ products/dashboards/mcp/tools.yaml
@@ -406,10 +406,10 @@ tools:
         description: >
             Run live queries for dashboard widget tiles and return results. Requires the dashboard-widgets feature flag.
             Call dashboard-get first to find widget tile IDs (tiles with a widget field). Pass tile_ids as a
-            comma-separated list.             Each successful result includes results (rows), limit (page size),
-            totalCount, totalCountCapped (when the true total exceeds the count cap), and hasMore. Returns per-tile
-            errors when the caller lacks product access for a widget type (see required_product_access on
-            dashboard-widget-catalog-list). Not available on public/shared dashboard views.
+            comma-separated list. Each successful result is a widget-type-specific payload (see run_widgets OpenAPI
+            response and dashboard-widget-catalog-list config_schema_hints). Returns per-tile errors when the caller
+            lacks product access for a widget type (see required_product_access on dashboard-widget-catalog-list). Not
+            available on public/shared dashboard views.
         param_overrides:
             id:
                 cast: string-int
